utils: functions for loading jsons and dictionaries as objects
