+++ renderapi/utils.py
@@ -24,3 +24,14 @@ def stripLogger(logger_tostrip):
     if logger_tostrip.handlers:
         for handler in logger_tostrip.handlers:
             logger_tostrip.removeHandler(handler)
+
+
+def _load_dict(obj, d):
+    obj.__dict__.update({k: v for k, v in d.items()})
+
+
+def _load_json(obj, j):
+    '''load object from dictionary-style json'''
+    with open(j, 'r') as f:
+        jd = json.load(f)
+    _load_dict(obj, jd)
